refactor(auth): consolidate useSession into a single SessionProvider
Multiple components each called Better Auth's useSession() independently,
creating duplicate subscriptions to the session refresh manager and
causing excessive get-session API requests that triggered 429 rate limits.
