+++ components/Header/Header.tsx
@@ -4,7 +4,8 @@ import React, { useCallback, useRef, useState } from 'react';
 import styles from './Header.module.css';
 import Link from 'next/link';
 import Image from 'next/image';
-import { signOut, useSession } from '@/lib/auth-client';
+import { signOut } from '@/lib/auth-client';
+import { useSession } from '@/context/SessionProvider';
 import { usePathname, useRouter } from 'next/navigation';
 import { useClickOutside, useEscapeKey, useFocusTrap } from '@/lib/hooks';
 import { useFeedback } from '@/context/FeedbackProvider';
